+++ src/ index.ts
@@ -1,0 +1,2 @@
+export { RedisCoreError } from "./redis.error"
+export { RedisConnector } from "./redis.service"