+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v1.0.0] 2024-04-25
+## [v1.0.0] 2024-05-07
 ### Added
+* Redis Connector Service
+* Redis simple methods: SET, GET, DEL, HSET, HGET, HDEL